services/model-orchestration/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn==0.25.0
pydantic==2.5.3
pydantic[email]==2.5.3  # Includes email-validator
email-validator==2.1.0  # Explicitly include for clarity
starlette==0.36.3
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.21.0  # For testing with SQLite

# Messaging
redis==5.0.1
aioredis==2.0.1

# AI Model Providers
openai==1.10.0
anthropic==0.8.0
httpx==0.26.0
tiktoken==0.5.2

# Monitoring and logging
prometheus-client==0.19.0
sentry-sdk==1.39.1
structlog==23.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0
# TestClient is included in FastAPI

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
pydantic-settings==2.1.0  # Required for BaseSettings
regex>=2022.1.18  # Required by tiktoken
